requirements.local.txt
absl-py==2.3.1
accelerate==1.9.0
AIAK-Training-LLM @ file:///share/data/drive_3/mobile_vlm/LLaVA-OneVision-1.5
aiobotocore==2.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anykeystore==0.2
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
botocore==1.41.5
braceexpand==0.1.7
build==1.3.0
cfgv==3.5.0
click==8.3.1
cryptacular==1.6.2
datasets==2.19.2
defusedxml==0.7.1
dill==0.3.8
distlib==0.4.0
einops==0.6.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2024.3.1
ftfy==6.1.3
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
grpcio==1.76.0
hf-xet==1.2.0
huggingface-hub==0.36.0
hupper==1.12.1
hydra-core==1.3.2
identify==2.6.15
importlib_metadata==8.7.0
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
Levenshtein==0.27.1
Markdown==3.10
megatron-energon==5.0.0
mkl-service==2.5.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.3
ninja==1.13.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
packaging==25.0
pandarallel==1.6.5
pandas==2.3.3
PasteDeploy==3.1.0
pbkdf2==1.3
pillow==11.3.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.5.1
pre-commit==3.2.2
propcache==0.4.1
protobuf==3.20.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyarrow-hotfix==0.7
pybind11==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pyproject_hooks==1.2.0
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python3-openid==3.2.0
pytz==2025.2
qwen-vl-utils==0.0.14
RapidFuzz==3.14.2
regex==2025.11.3
repoze.sendmail==4.4.1
requests-oauthlib==2.0.0
s3fs==2024.3.1
safetensors==0.7.0
sentencepiece==0.1.98
sentry-sdk==2.47.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tiktoken==0.5.2
timm==1.0.3
tokenizers==0.21.4
tomli==2.3.0
tqdm==4.67.1
transaction==5.0
transformers==4.53.1
translationstring==1.4
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
velruse==1.1.1
venusian==3.1.1
virtualenv==20.35.4
wandb==0.21.0
wcwidth==0.2.14
webdataset==1.0.2
WebOb==1.8.9
Werkzeug==3.1.4
wrapt==1.16.0
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.deprecation==6.0
zope.interface==8.1.1
zope.sqlalchemy==4.1
